Mount all iso listed in bootstrap config file
Add a products entry in /etc/metalk8s/bootstrap.yaml which will contain
the product path or the product iso. If product path is provided,
expose saltstack file_roots and pillar_roots
into salt master configuration. If a iso image is provided, mount the iso
first before performing the same operation.

As set of assertion is also made to verify the product version. In this
case, if the products:metalk8s field contains an iso, isoinfo binary is
required to extract the product version

Relates: #924